==> R_scripts/4-3.summary_biol_func_driver_genes.Rmd <==
---
title: "SC20054 - Transcriptomic signatures of iPSC-derived dopamine neurons"
subtitle: "Summary biological function overrepresentatn of cluster-specific driver genes of velocity estimation with scVelo"
author: Stephanie Strohbuecker
output: 
  html_document:
    df_print: paged
---

# Set up

```{r}
RMD_file_name <- "4-3.summary_biol_func_driver_genes"
```


## Libraries

```{r}
library(tidyverse)
library(here)

library(ggpubr)

library(clusterProfiler)
```

## File locations
Specify the location of input and output
```{r}
# Specify the location of input and output
Reactome <-
  read_rds(
    here(
      "output",
      "4-2.Reactome_enrichment_driver_genes",
      "Driver_genes_top100_Reactome_enrichment.RDS"
    )
  )

GO_BP <-
  read_rds(
    here(
      "output",
      "4-1.GO_enrichment_driver_genes",
      "Driver_genes_top100_BPGO_enrichment.RDS"
    )
  )

# Path to output directory
output <- str_c(here(), "output", RMD_file_name, sep = "/")

if(!dir.exists(output)){
  dir.create(output, recursive = TRUE)
}

figures_out <- str_c(here(), "figures", RMD_file_name, sep = "/")

if(!dir.exists(figures_out)){
  dir.create(figures_out, recursive = TRUE)
}

```

```{r}
Reactome <- flatten(Reactome)
GO_BP <- flatten(GO_BP)
```

```{r fig.height=14, fig.width=8, message=FALSE, warning=FALSE}
(biol_func <- map(names(Reactome), function(cluster_id){
  
ggarrange(
  Reactome[[cluster_id]]$dotplot_Reactome +
    theme(legend.position = "bottom",
          legend.box = "vertical") +
    guides(
      colour = guide_colourbar(
        order = 1,
        barwidth = 10,
        title.position = "top",
        label.theme = element_text(
          angle = 90,
          vjust = 0.5,
          size = 10
        )
      ),
      size = guide_legend(order = 2)
    ) +
    scale_y_discrete(
      label = function(x)
        if_else(
          str_length(x) > 60 ,
                  str_wrap(x, 60),
          x
        )
    ),
  GO_BP[[cluster_id]]$dotplot_GO_BP +
    theme(legend.position = "bottom",
          legend.box = "vertical",
          legend.margin = margin(c(0, 0, 0, 0))) +
    guides(
      colour = guide_colourbar(
        order = 1,
        barwidth = 10,
        title.position = "top",
        label.theme = element_text(
          angle = 90,
          vjust = 0.5,
          size = 10
        )
      ),
      size = guide_legend(order = 2)
    ) +
    scale_y_discrete(
      label = function(x)
        if_else(
          str_length(x) > 60 ,
                  str_wrap(x, 60),
          x
        )
    ),
  ncol = 1,
  nrow = 2,
  align = "hv"
) %>%
  annotate_figure(
    top = text_grob(
      str_c("Biological function enrichment for\ncluster", cluster_id, sep = " "),
      face = "bold"
    )
  )}))
```

```{r}
pdf(str_c(figures_out, "Biological_function_overrepresentation_top100_cluster_specific_driver_genes.pdf", sep = "/"), onefile=TRUE, height = 14, width = 8)
biol_func
dev.off()
```


```{r}
openxlsx::write.xlsx(
  c(
    map(Reactome, ~ .x$Reactome_sig_res) %>%
      set_names(str_c(names(.), "Reactome", sep = "_")),
    map(GO_BP, ~ .x$BP_GO_sig_res) %>%
      set_names(str_c(names(.), "GP_BP", sep = "_"))
  ),
  file = str_c(
    output,
    "Biological_function_overrepresentation_top100_cluster_specific_driver_genes.xlsx",
    sep = "/"
  )
)
```
